# apiary-OpenApi3Json.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is the documentation for the [AHEM Ad Hoc Disposable Temporary Email service API](https://www.ahem.email).  AHEM is a free temp mail service that allows recieving of ad-hoc emails to a mailbox of your choice. No  registration or any other preparation is needed. Just send an email, and it's there.<br> Its simple and intuitive api makes it a perfect tool for testing and automation of any product that use email.",
    "version": "1.0.0",
    "title": "AHEM - Ad Hoc Email Service",
    "termsOfService": "https://www.ahem.email/privacy",
    "contact": {}
  },
  "servers": [
    {
      "url": "https://www.ahem.email/api/",
      "variables": {}
    }
  ],
  "paths": {
    "/alive": {
      "get": {
        "summary": "Get system check results",
        "description": "",
        "operationId": "GetSystemCheckResults",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "api": true,
                    "smtp": true,
                    "db": true
                  },
                  "$ref": "#/components/schemas/GetSystemCheckResultsResponse"
                },
                "example": {
                  "success": true,
                  "api": true,
                  "smtp": true,
                  "db": true
                }
              }
            }
          }
        },
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      }
    },
    "/properties": {
      "get": {
        "summary": "Get server properties",
        "description": "",
        "operationId": "GetServerProperties",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "serverBaseUri": "https://www.ahem.email",
                    "emailDeleteInterval": 86400,
                    "emailDeleteAge": 7200,
                    "allowedDomains": [
                      "ahem.email",
                      "mail.ahem.email",
                      "ahem-email.com"
                    ],
                    "allowAutocomplete": false,
                    "customText": null
                  },
                  "$ref": "#/components/schemas/GetServerPropertiesResponse"
                },
                "example": {
                  "serverBaseUri": "https://www.ahem.email",
                  "emailDeleteInterval": 86400,
                  "emailDeleteAge": 7200,
                  "allowedDomains": [
                    "ahem.email",
                    "mail.ahem.email",
                    "ahem-email.com"
                  ],
                  "allowAutocomplete": false,
                  "customText": null
                }
              }
            }
          }
        },
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      }
    },
    "/mailbox/{mailbox}/email": {
      "get": {
        "summary": "List All Emails in a mailbox",
        "description": "",
        "operationId": "ListAllEmailsInAnMailBox",
        "security": "bearerAuth",
        "parameters": [
          {
            "name": "mailboxName",
            "in": "path",
            "description": "mailbox name.",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "mailboxName"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "example": [
                      {
                        "emailId": "5bc3c17cb76d8f4e1099725c",
                        "sender": {
                          "address": "test@my.domain.com",
                          "name": "AHEM Test!"
                        },
                        "subject": "AHEM mail test!",
                        "timestamp": 1539555708477,
                        "isRead": true
                      }
                    ],
                    "$ref": "#/components/schemas/ListAllEmailsInAnmailboxResponse"
                  },
                  "description": ""
                },
                "example": [
                  {
                    "emailId": "5bc3c17cb76d8f4e1099725c",
                    "sender": {
                      "address": "test@my.domain.com",
                      "name": "AHEM Test!"
                    },
                    "subject": "AHEM mail test!",
                    "timestamp": 1539555708477,
                    "isRead": true
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unexpected error in API call. See HTTP response body for details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAllEmailsInAnmailbox401Response"
                },
                "example": {
                  "success": false,
                  "message": "No token provided."
                }
              }
            }
          },
          "404": {
            "description": "Unexpected error in API call. See HTTP response body for details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAllEmailsInAnmailbox404Response"
                },
                "example": {
                  "error": "mailbox IS EMPTY!"
                }
              }
            }
          }
        },
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "GetSystemCheckResultsResponse": {
        "title": "Get system check results response",
        "required": [
          "success",
          "api",
          "smtp",
          "db"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "api": {
            "type": "boolean",
            "example": true
          },
          "smtp": {
            "type": "boolean",
            "example": true
          },
          "db": {
            "type": "boolean",
            "example": true
          }
        },
        "example": {
          "success": true,
          "api": true,
          "smtp": true,
          "db": true
        }
      },
      "GetServerPropertiesResponse": {
        "title": "Get server properties response",
        "required": [
          "serverBaseUri",
          "emailDeleteInterval",
          "emailDeleteAge",
          "allowedDomains",
          "allowAutocomplete",
          "customText"
        ],
        "type": "object",
        "properties": {
          "serverBaseUri": {
            "type": "string",
            "example": "https://www.ahem.email"
          },
          "emailDeleteInterval": {
            "type": "integer",
            "format": "int32",
            "example": 86400
          },
          "emailDeleteAge": {
            "type": "integer",
            "format": "int32",
            "example": 7200
          },
          "allowedDomains": {
            "type": "array",
            "items": {
              "type": "string",
              "example": [
                "ahem.email",
                "mail.ahem.email",
                "ahem-email.com"
              ]
            },
            "description": ""
          },
          "allowAutocomplete": {
            "type": "boolean",
            "example": false
          },
          "customText": {
            "type": "string",
            "nullable": true
          }
        },
        "example": {
          "serverBaseUri": "https://www.ahem.email",
          "emailDeleteInterval": 86400,
          "emailDeleteAge": 7200,
          "allowedDomains": [
            "ahem.email",
            "mail.ahem.email",
            "ahem-email.com"
          ],
          "allowAutocomplete": false,
          "customText": null
        }
      },
      "ListAllEmailsInAnmailboxResponse": {
        "title": "List All Emails in an mailbox response",
        "required": [
          "emailId",
          "sender",
          "subject",
          "timestamp",
          "isRead"
        ],
        "type": "object",
        "properties": {
          "emailId": {
            "type": "string",
            "example": "5bc3c17cb76d8f4e1099725c"
          },
          "sender": {
            "type": "object",
            "example": {
              "address": "test@my.domain.com",
              "name": "AHEM Test!"
            }
          },
          "subject": {
            "type": "string",
            "example": "AHEM mail test!"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "example": 1539555708477
          },
          "isRead": {
            "type": "boolean",
            "example": true
          }
        },
        "example": {
          "emailId": "5bc3c17cb76d8f4e1099725c",
          "sender": {
            "address": "test@my.domain.com",
            "name": "AHEM Test!"
          },
          "subject": "AHEM mail test!",
          "timestamp": 1539555708477,
          "isRead": true
        }
      },
      "ListAllEmailsInAnmailbox404Response": {
        "title": "List All Emails in an mailbox 404 response",
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "mailbox IS EMPTY!"
          }
        },
        "example": {
          "error": "mailbox IS EMPTY!"
        }
      },
      "ListAllEmailsInAnmailbox401Response": {
        "title": "List All Emails in an mailbox 401 response",
        "required": [
          "success",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "No token provided."
          }
        },
        "example": {
          "success": false,
          "message": "No token provided."
        }
      }
    }
  },
  "security": [
    {}
  ],
  "tags": []
}